GstSHM: Initial import.
What we are building is a pair of producer and consumer that link
various melt processes via shm using the same control semantics
as the shmsink / shmsrc GStreamer elements.

The on memory format is not compatible with GStreamer but elements
to mux and demux it will come soon.

See README for more details.